requirements.txt
streamlit
requests
beautifulsoup4
pillow
vaderSentiment
